+++ src/components/Pagination/Pagination.tsx
@@ -18,6 +18,7 @@ export interface PaginationProps {
 
 const Pagination: React.FC<PaginationProps> = props => {
   const { pageSize, pageIndex = 0, nextPage, pageCount, itemCount, gotoPage, pageCountClamp = 5 } = props
+
   if (pageIndex >= pageCount) {
     // eslint-disable-next-line no-console
     console.warn(`Pagination: pageIndex (${pageIndex}) can't be more than/equal to pageCount (${pageCount})`)
@@ -27,6 +28,28 @@ const Pagination: React.FC<PaginationProps> = props => {
     // eslint-disable-next-line no-console
     console.warn(
       `Pagination: pageSize (${pageSize}) * pageCount (${pageCount}) can't be less than itemCount (${itemCount})`
+    )
+  }
+  if (pageCountClamp < 5) {
+    // eslint-disable-next-line no-console
+    console.warn(`Pagination: pageCountClamp (${pageCountClamp}) can't be less than 5`)
+    return null
+  }
+
+  const moreLeft = pageIndex + 1 > pageCountClamp
+  const moreRight = pageCount - pageIndex > pageCountClamp
+
+  const renderPage = (pageNumber: number) => {
+    const isCurrent = pageIndex + 1 === pageNumber
+    return (
+      <Button
+        key={`page-${pageNumber}`}
+        minimal={!isCurrent}
+        text={pageNumber}
+        intent={isCurrent ? 'primary' : 'none'}
+        disabled={isCurrent}
+        onClick={() => gotoPage(pageNumber - 1)}
+      />
     )
   }
 
@@ -49,7 +72,7 @@ const Pagination: React.FC<PaginationProps> = props => {
       )}
       {itemCount > pageSize ? (
         <Layout.Horizontal spacing="small">
-          {pageCount < pageCountClamp + 3 ? (
+          {pageCount <= pageCountClamp * 2 ? (
             Array.from(Array(pageCount).keys()).map(pageNumber => {
               const isCurrent = pageIndex === pageNumber
               return (
@@ -64,47 +87,50 @@ const Pagination: React.FC<PaginationProps> = props => {
               )
             })
           ) : (
-            <>
-              {Array.from(Array(pageCountClamp).keys()).map(pageNumber => {
-                const isCurrent = pageIndex === pageNumber
-                return (
-                  <Button
-                    key={`page-${pageNumber}`}
-                    minimal={!isCurrent}
-                    text={pageNumber + 1}
-                    intent={isCurrent ? 'primary' : 'none'}
-                    disabled={isCurrent}
-                    onClick={() => gotoPage(pageNumber)}
-                  />
-                )
-              })}
-              <SelectV2
-                items={Array.from(Array(pageCount - pageCountClamp - 1).keys()).map(pageNumber => {
-                  return {
-                    label: `${pageNumber + pageCountClamp + 1}`,
-                    value: pageNumber + pageCountClamp
-                  }
-                })}
-                onChange={item => gotoPage(item.value as number)}
-                filterable={false}>
-                <Button icon="more" minimal />
-              </SelectV2>
-              {/* Special treatment for last button */}
-              <Button
-                minimal={pageIndex != pageCount - 1}
-                text={pageCount}
-                intent={pageIndex === pageCount - 1 ? 'primary' : 'none'}
-                disabled={pageIndex === pageCount - 1}
-                onClick={() => gotoPage(pageCount - 1)}
-              />
-            </>
+            <React.Fragment>
+              {Array.from({ length: moreLeft ? 2 : pageCountClamp }, (_, i) => renderPage(i + 1))}
+              {moreLeft && (
+                <SelectV2
+                  items={Array.from(Array(moreRight ? pageIndex - 3 : pageCount - pageCountClamp - 2).keys()).map(
+                    pageNumber => {
+                      return {
+                        label: `${pageNumber + 3}`,
+                        value: pageNumber + 2
+                      }
+                    }
+                  )}
+                  onChange={item => gotoPage(item.value as number)}
+                  filterable={false}>
+                  <Button icon="more" minimal width="100%" />
+                </SelectV2>
+              )}
+              {moreLeft && moreRight && Array.from({ length: 3 }, (_, i) => renderPage(pageIndex + i))}
+              {moreRight && (
+                <SelectV2
+                  items={Array.from(
+                    Array(moreLeft ? pageCount - pageIndex - 4 : pageCount - pageCountClamp - 2).keys()
+                  ).map(pageNumber => {
+                    return {
+                      label: `${pageNumber + (moreLeft ? pageIndex + 3 : pageCountClamp + 1)}`,
+                      value: pageNumber + (moreLeft ? pageIndex + 2 : pageCountClamp)
+                    }
+                  })}
+                  onChange={item => gotoPage(item.value as number)}
+                  filterable={false}>
+                  <Button icon="more" minimal width="100%" />
+                </SelectV2>
+              )}
+              {moreRight
+                ? [pageCount - 1, pageCount].map(i => renderPage(i))
+                : Array.from({ length: pageCountClamp }, (_, i) => renderPage(i + 1 + pageCount - pageCountClamp))}
+            </React.Fragment>
           )}
           {pageCount > pageCountClamp && nextPage ? (
             <Button
               minimal
               icon="chevron-right"
               onClick={() => nextPage(pageIndex + 1)}
-              disabled={pageIndex === pageCount}
+              disabled={pageIndex + 1 === pageCount}
               style={{ marginRight: 'var(--spacing-medium)' }}
             />
           ) : null}